refactor: dry code per @shon-button's recommendations
